pyproject.toml
[tool.poetry]
name = "gatekeeper-bot"
version = "0.1.0"
description = ""
authors = ["Pármenas Haniel <parmenashp@gmail.com>"]
readme = "README.md"
packages = [{include = "gatekeeper_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
asyncpg = "^0.27.0"
pyjwt = "^2.6.0"
loguru = "^0.6.0"
jishaku = "^2.5.1"
fluent-runtime = "^0.3.1"

[tool.black]
line-length = 119

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
